*: refactor the testing harness to use legate.core's testing harness
