--- tests/_utils/__init__.py
@@ -1,74 +1,0 @@
-# Copyright 2022 NVIDIA Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-"""Utilities and helpers for implementing the Cunumeric custom test runner.
-
-"""
-from __future__ import annotations
-
-from typing import Union
-from typing_extensions import Literal, TypeAlias
-
-FeatureType: TypeAlias = Union[
-    Literal["cpus"], Literal["openmp"], Literal["cuda"],
-]
-
-#: Value to use if --cpus is not specified.
-DEFAULT_CPUS_PER_NODE = 4
-
-#: Value to use if --gpus is not specified.
-DEFAULT_GPUS_PER_NODE = 1
-
-# Value to use if --fbmem is not specified
-DEFAULT_GPU_MEMORY_BUDGET = 6 << 30
-
-# TODO: (bryevdv) configurable?
-DEFAULT_GPU_PARALLELISM = 16
-
-#: Value to use if --omps is not specified.
-DEFAULT_OMPS_PER_NODE = 1
-
-#: Value to use if --ompthreads is not specified.
-DEFAULT_OMPTHREADS = 4
-
-#: Default values to apply to normalize the testing environment.
-DEFAULT_PROCESS_ENV = {
-    "LEGATE_TEST": "1",
-    "REALM_SYNTHETIC_CORE_MAP": "",
-}
-
-#: Width for terminal ouput headers and footers.
-UI_WIDTH = 60
-
-#: Feature values that are accepted for --use, in the relative order
-#: that the corresponding test stages should always execute in
-FEATURES: tuple[FeatureType, ...] = (
-    "cpus",
-    "openmp",
-    "cuda",
-)
-
-#: Paths to example files that should be skipped.
-SKIPPED_EXAMPLES = {
-    "examples/reference_amg.py",
-    "examples/plot.py",
-    "examples/pyamg_legate_test.py",
-    "examples/pde-petsc.py",
-}
-
-#: Extra arguments to supply when specific examples are executed.
-PER_FILE_ARGS = {
-    "examples/lstm_full.py": ["--file", "resources/lstm_input.txt"],
-    "examples/amg.py": ["-nodes", "128"],
-}
