+++ tests/integration/test_dia.py
