+++ tests/integration/test_solvers.py
