+++ tests/integration/test_spatial.py
